requirements.txt
PySide6>=6.10.1
requests>=2.32.0
psutil>=5.9.0
python-dotenv>=1.0.0
